Remove unused function & separate certain models
