+++ models/src/user/mod.rs
@@ -1,9 +1,10 @@
-use bitflags::bitflags;
-use postgres_types::{FromSql, Type, ToSql, IsNull, to_sql_checked};
-use bytes::BytesMut;
-use serde::{Deserialize, Serialize, Deserializer, Serializer};
+mod flags;
+
+use serde::{Deserialize, Serialize};
 
 use crate::FromRow;
+
+pub use flags::UserFlags;
 
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
 pub struct RoUser {
@@ -31,15 +32,7 @@ pub struct QueueUser {
     pub verified: bool
 }
 
-bitflags! {
-    pub struct UserFlags: i64 {
-        const NONE = 0;
-        const ALPHA = 1;
-        const BETA = 1 << 1;
-        const STAFF = 1 << 2;
-        const PARTNER = 1 << 3;
-    }
-}
+
 
 impl FromRow for RoUser {
     fn from_row(row: tokio_postgres::Row) -> Result<Self, tokio_postgres::Error> {
@@ -90,40 +83,5 @@ impl FromRow for QueueUser {
             roblox_id,
             verified
         })
-    }
-}
-
-impl<'a> FromSql<'a> for UserFlags {
-    fn from_sql(ty: &Type, raw: &'a [u8]) -> Result<Self, Box<dyn std::error::Error + Sync + Send>> {
-        let bits = i64::from_sql(ty, raw)?;
-        Ok(Self::from_bits_truncate(bits))
-    }
-
-    fn accepts(ty: &Type) -> bool {
-        <i64 as FromSql>::accepts(ty)
-    }
-}
-
-impl ToSql for UserFlags {
-    fn to_sql(&self, ty: &Type, out: &mut BytesMut) -> Result<IsNull, Box<dyn std::error::Error + Sync + Send>> {
-        i64::to_sql(&self.bits, ty, out)
-    }
-
-    fn accepts(ty: &Type) -> bool {
-        <i64 as ToSql>::accepts(ty)
-    }
-
-    to_sql_checked!();
-}
-
-impl<'de> Deserialize<'de> for UserFlags {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        Ok(Self::from_bits_truncate(i64::deserialize(deserializer)?))
-    }
-}
-
-impl Serialize for UserFlags {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        serializer.serialize_i64(self.bits())
     }
 }
